Restructure of translation code ahead of auth changes
