+++ packages/graphql/src/translate/create-projection-and-params.ts
@@ -22,7 +22,7 @@ import { mergeDeep } from "@graphql-tools/utils";
 import Cypher from "@neo4j/cypher-builder";
 import type { Node } from "../classes";
 import type { GraphQLOptionsArg, GraphQLWhereArg, Context, GraphQLSortArg, CypherFieldReferenceMap } from "../types";
-import { createAuthPredicates } from "./create-auth-and-params";
+import { createAuthPredicates } from "./create-auth-predicates";
 import { createDatetimeExpression } from "./projection/elements/create-datetime-element";
 import { createPointExpression } from "./projection/elements/create-point-element";
 import mapToDbProperty from "../utils/map-to-db-property";
@@ -94,7 +94,7 @@ export default function createProjectionAndParams({
                     operations: "READ",
                     context,
                     allow: {
-                        parentNode: node,
+                        node,
                         varName,
                     },
                 });
@@ -188,9 +188,7 @@ export default function createProjectionAndParams({
                         // The nested projection will be surrounded by brackets, so we want to remove
                         // any linebreaks, and then the first opening and the last closing bracket of the line,
                         // as well as any surrounding whitespace.
-                        const nestedProj = recurse.projection
-                            .getCypher(env)
-                            .replaceAll(/(^\s*{\s*)|(\s*}\s*$)/g, "");
+                        const nestedProj = recurse.projection.getCypher(env).replaceAll(/(^\s*{\s*)|(\s*}\s*$)/g, "");
 
                         return `{ __resolveType: "${refNode.name}", __id: id(${varName.getCypher(env)})${
                             nestedProj && `, ${nestedProj}`
